refactor: rename model-preview-selector to main-model-selector
