+++ src/components/specific/models/models-card/main-model-selector/MainModelSelector.scss
@@ -1,4 +1,4 @@
-.model-preview-selector {
+.main-model-selector {
   position: relative;
   width: 75%;
   height: 75%;
@@ -33,7 +33,7 @@
     flex-grow: 1;
     overflow-y: auto;
 
-    .model-preview-item {
+    .model-item {
       position: relative;
       width: 215px;
       height: 247px;
