+++ src/components/specific/models/models-table/model-name-cell/ModelNameCell.scss
