+++ src/components/specific/models/models-table/model-name-cell/ModelNameCell.vue
@@ -53,9 +53,9 @@
 
 <script>
 import { ref, watch } from "vue";
-import { MODEL_CONFIG, MODEL_TYPE } from "../../../../../../config/models.js";
-import { useModels } from "../../../../../../state/models.js";
-import { debounce } from "../../../../../../utils/async.js";
+import { MODEL_CONFIG, MODEL_TYPE } from "../../../../../config/models.js";
+import { useModels } from "../../../../../state/models.js";
+import { debounce } from "../../../../../utils/async.js";
 
 export default {
   props: {
