+++ src/components/specific/models/models-table/model-status-cell/ModelStatusCell.scss
